+++ CHANGELOG.md
@@ -7,6 +7,8 @@ with regard to its JavaScript and TypeScript bindings' APIs and the Ethereum ABI
 
 ## [Unreleased]
 
+- Replaced vesting contract with one adapted from Melonport
+
 ## [0.1.0] - 2020-12-01
 ### Added
 - The initial version of the funding pool contract